Create Python Shopping app
This is a shopping app or program coded in python programming language. It is still in very early beta. It is version 0.2 and i hope to add more in the upcoming versions. I have more projects i am working on off of this so if you want to help just request it.